+++ testutils_robolectric/src/main/java/com/google/android/exoplayer2/testutil/StubExoPlayer.java
@@ -150,6 +150,26 @@ public abstract class StubExoPlayer implements ExoPlayer {
   }
 
   @Override
+  public boolean hasPrevious() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public void previous() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public boolean hasNext() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public void next() {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
   public void setPlaybackParameters(PlaybackParameters playbackParameters) {
     throw new UnsupportedOperationException();
   }